repos created

+++ src/main/java/org/openapitools/persistence/repositories/DocumentRepository.java
@@ -3,5 +3,5 @@ package org.openapitools.persistence.repositories;
 import org.openapitools.persistence.entities.DocumentEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DocumentRepository extends JpaRepository<DocumentEntity, Long> {
+public interface DocumentRepository  extends JpaRepository<DocumentEntity, Integer> {
 }
